requirements.txt
numpy
PyYAML
pydantic
pytest
tqdm
datasets
httpx